NSArray index type is NSUInteger, not long
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105740
Tested-by: Jenkins

+++ fpicker/source/aqua/SalAquaFolderPicker.mm
@@ -114,7 +114,7 @@ OUString SAL_CALL SalAquaFolderPicker::getDirectory()
         files = [static_cast<NSOpenPanel*>(m_pDialog) URLs];
     }
 
-    long nFiles = [files count];
+    NSUInteger nFiles = [files count];
     SAL_INFO("fpicker.aqua", "# of items: " << nFiles);
 
     if (nFiles < 1) {
